+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 - Smart Loading (reuse if already exists)
 - do_save (embed packaging in actions)
-- load_folder
+- pkg_all #folder
 
 
 0.3.0 (2022-02-28)